// src/entities/movie/model/response.ts
import { Content, ProfileContent } from '@/entities/movie/model/content'
import { Genre } from '@/entities/movie/model/genre'
import { Backdrop, Logo, Poster } from '@/entities/movie/model/image'
import { Movie, MyLikeMovie } from '@/entities/movie/model/movie'

interface Dates {
  /** 가장 최근에 등록된 상영 날짜 */
  maximum: string
  /** 가장 마지막에 등록된 사영 날짜 */
  minimum: string
}

export type Pagination<T> = {
  results: T[]
  page: number
  total_pages: number
  total_results: number
  /** 검색 기간 */
  dates?: Dates
}

interface BelongsToCollection {
  id: number
  name: string
  poster_path: string
  backdrop_path: string
}

interface ProductionCompany {
  id: number
  logo_path: string
  name: string
  origin_country: string
}

interface ProductionCountry {
  iso_3166_1: string
  name: string
}

interface SpokenLanguage {
  english_name: string
  iso_639_1: string
  name: string
}

export type MoviesResponse = Pagination<Movie>

export interface DetailResponse {
  adult: boolean
  backdrop_path: string
  belongs_to_collection?: BelongsToCollection
  budget: number
  genres: Genre[]
  homepage: string
  id: number
  imdb_id: string
  origin_country: string[]
  original_language: string
  original_title: string
  overview: string
  popularity: number
  poster_path: string
  production_companies: ProductionCompany[]
  production_countries: ProductionCountry[]
  release_date: string
  revenue: number
  runtime: number
  spoken_languages: SpokenLanguage[]
  status: string
  tagline: string
  title: string
  video: boolean
  vote_average: number
  vote_count: number
}

export interface ImageResponse {
  backdrops: Backdrop[]
  id: number
  logos: Logo[]
  posters: Poster[]
}

export type ContentResponse = Pagination<Content>

export type MyBookMarkResponse = Pagination<MyLikeMovie>

export type MyReviewResponse = Pagination<ProfileContent>